+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Random CHANGELOG
 
+### Version 1.3.0
+
+* new method ```chance```
+    * returns a boolean depending on the chance of the passed in number
+
 ### Version 1.2.0
 
 * new method ```keyFromObject```